1. global 패키지 재정의
