+++ src/main/java/com/nexters/phochak/domain/Likes.java
@@ -1,5 +1,6 @@
 package com.nexters.phochak.domain;
 
+import com.nexters.phochak.global.domain.BaseTime;
 import lombok.Builder;
 import lombok.Getter;
