+++ src/main/java/com/nexters/phochak/global/domain/BaseTime.java
@@ -1,4 +1,4 @@
-package com.nexters.phochak.domain;
+package com.nexters.phochak.global.domain;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Getter;
